extract out content to seperate files
