+++ website/content/features/code/code.ts
@@ -1,0 +1,3 @@
+export const tab1 = "Lambda function";
+export const tab2 = "Step function";
+export const tab3 = "Appsync resolver";
