+++ website/content/hero.ts
@@ -1,0 +1,10 @@
+export const versionRelease = "v1.02.5 Released";
+
+export const title = "Code the Cloud";
+
+export const body =
+  "Develop APIs, Functions, Workflows and Event-Driven Architectures with “Smart Cloud Resources” that have best practices built-in.  Least-privilege IAM Policies, environment variables, SDK clients and Domain Specific Languages are automatically configured.";
+
+export const docsLink = "Read Docs";
+
+export const githubLink = "Star us on GitUb";
