+++ website/content/site.ts
@@ -1,0 +1,1 @@
+export const githubUrl = "https://github.com/functionless/functionless";
